Add FrontEnd changes to enable deletion of document
Added button that will call the API to delete a document. It only
shows the button once the document has been processed to avoid any
issues while processing the embeddings.

However, still needs to automatically delete the Document from the list
as right now it requires a refresh on call documents.